+++ lib/stream.dart
@@ -19,6 +19,7 @@ import "package:rikulo_commons/browser.dart";
 import "package:rikulo_commons/logging.dart";
 
 import "plugin.dart";
+export "plugin.dart" show Router, DefaultRouter;
 import "src/version.dart" as version;
 
 export "package:rikulo_commons/browser.dart" show Browser;